flash-tix/notification-service/requirements.txt
fastapi
uvicorn
pika
python-dotenv
requests